requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
easyocr==1.7.2
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy<2
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic_core==2.27.2
pytesseract==0.3.13
python-bidi==0.6.3
PyYAML==6.0.2
scikit-image==0.25.0
scipy==1.15.1
shapely==2.0.6
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tesseract==0.1.3
tifffile==2025.1.10
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
uvicorn==0.34.0
